Add css changes for consistency (incomplete)

+++ src/Components/DateTime/FormatDate.jsx
@@ -1,7 +1,8 @@
 // KAT - DONE
 function FormatDate(dateString) {
     const date = new Date(dateString);
-    const month = date.toLocaleString('default', { month: 'long' });
+    // date.setHours(0, 0, 0, 0); To remove time from end of session date - but makes program detail date invalid - remove dependency
+    const month = date.toLocaleString('default', { month: 'short' });
     const day = date.getDate();
     const year = date.getFullYear();
     return `${day} ${month} ${year}`;
